Move `CORSMessage` to `utils/`
